requirements.txt
llama-index-core>=0.12.0
llama-index-llms-anthropic>=0.6.0
llama-index-embeddings-huggingface>=0.5.0
llama-index-vector-stores-chroma>=0.4.0
chromadb>=0.5.0
python-dotenv>=1.0.0
sentence-transformers>=3.0.0
mcp[cli]>=1.0.0
